refactor: replace logger, swagger in middleware folder, make cleaner run file
