+++ src/app.ts
@@ -1,11 +1,9 @@
-import express, { json, Express } from 'express';
-import { serve, setup } from 'swagger-ui-express';
+import express, { Express } from 'express';
 import passport from 'passport';
 
 import { sequelize } from '@config/dbConfig.js';
 import { appConfig } from '@config/appConfig.js';
 import { passportConfig } from '@config/pasportConfig.js';
-import swaggerSpec from '@utils/swaggerConf.js';
 
 import { setupMiddlewares } from '@middleware/_middlewares.js';
 import { setupRoutes } from '@routes/_routes.js';
@@ -15,9 +13,6 @@ import errorMiddleware from '@middleware/errorMiddleware.js';
 const app: Express = express();
 passportConfig(passport);
 
-app.use('/api-docs', serve, setup(swaggerSpec));
-app.use(json());
-
 setupMiddlewares(app);
 setupRoutes(app);
 
@@ -25,8 +20,8 @@ app.use(errorMiddleware);
 
 const connectDB = async () => {
     await sequelize.authenticate();
-    await sequelize.sync({ force: false });
     await setupAssociations();
+    await sequelize.sync({ force: false });
 
     console.log('db Connected');
 };
