+++ src/utils/logger.ts
@@ -1,4 +1,3 @@
-import { Request, Response, NextFunction } from 'express';
 import pino from 'pino';
 
 const logger = pino({
@@ -17,44 +16,4 @@ const logger = pino({
     },
 });
 
-const loggerMiddleware = async (
-    req: Request,
-    res: Response,
-    next: NextFunction,
-): Promise<void> => {
-    const start = Date.now();
-
-    res.on('finish', () => {
-        const duration = Date.now() - start;
-
-        if (res.statusCode >= 500) {
-            logger.error({
-                method: req.method,
-                url: req.originalUrl,
-                statusCode: res.statusCode,
-                responseTime: duration,
-                remoteAddress: req.socket.remoteAddress,
-            });
-        } else if (res.statusCode >= 400) {
-            logger.warn({
-                method: req.method,
-                url: req.originalUrl,
-                statusCode: res.statusCode,
-                responseTime: duration,
-                remoteAddress: req.socket.remoteAddress,
-            });
-        } else {
-            logger.info({
-                method: req.method,
-                url: req.originalUrl,
-                statusCode: res.statusCode,
-                responseTime: duration,
-                remoteAddress: req.socket.remoteAddress,
-            });
-        }
-    });
-
-    next();
-};
-
-export { logger, loggerMiddleware };
+export { logger };
