+++ src/utils/swaggerConf.ts
@@ -1,5 +1,3 @@
-import swaggerJSDoc from 'swagger-jsdoc';
-
 const swaggerOptions = {
     definition: {
         openapi: '3.0.0',
@@ -18,6 +16,4 @@ const swaggerOptions = {
     apis: ['./controllers/*.ts'],
 };
 
-const swaggerSpec = swaggerJSDoc(swaggerOptions);
-
-export default swaggerSpec;
+export default swaggerOptions;
